+++ src/opentype/color/svg.rs
@@ -57,7 +57,7 @@ impl SVG {
     pub(crate) fn get_svg(
         &self,
         gid: u32,
-        fonsize: f32,
+        fonsize: f64,
         fontunit: &str,
         layout: &crate::fontreader::HorizontalLayout,
         _: f64,
